Improved restarting a stalled app
Restarting the app every day was a blunt tool for fixing potentially
unresponsive apps. This commit adds a check for a file that is updated
every time the photo app checks in with the host. If there is more than
an hour without the photo app checking in, the app is restarted.

This also adds the ability to pause the photoctl script. When debugging,
it's nice to be able to turn off the cron job without having to mess
with crontab. By adding a third state for photoctl, "PAUSE", the
automtic stop/start of the app can be turned off as needed with
togglepause.py.

+++ photoctl.py
@@ -4,6 +4,8 @@ from __future__ import print_function
 import datetime
 import os
 from subprocess import Popen, PIPE
+import time
+
 
 HOMEDIR = "/home/pi/projects/photoviewer"
 DATE_FMT = "%Y-%m-%d %H:%M:%S"
@@ -18,6 +20,11 @@ def runproc(cmd, wait=True):
         stdout_text, stderr_text = proc.communicate()
         return stdout_text, stderr_text
 
+
+def killapp(pid):
+    os.kill(int(photoapp_out), 15)
+
+
 now = datetime.datetime.now().strftime(DATE_FMT)
 PHOTO_PID_FILE = "%s/photo.pid" % HOMEDIR
 with open(PHOTO_PID_FILE) as ff:
@@ -27,14 +34,30 @@ PHOTOAPP_CMD = "cd %s; python3 photo.py &" % HOMEDIR
 PID_CMD = "ps --no-headers -p %s" % PHOTO_PID
 photoapp_out, err = runproc(PID_CMD)
 photoapp_running = bool(photoapp_out)
+MAX_HOST_CHECK_INTERVAL = 3600
+HOST_CHECKED_FILE = "%s/.host_checked" % HOMEDIR
+try:
+    mtime = os.stat(HOST_CHECKED_FILE).st_mtime
+except FileNotFoundError:
+    # Running with an older photo app that doesn't create the file.
+    mtime = time.time()
+interval = time.time() - mtime
+restart = interval > MAX_HOST_CHECK_INTERVAL
 
 with open("/home/pi/.photoctl") as ff:
     action = ff.read().strip()
+if action == "PAUSE":
+    # Don't make any changes
+    exit()
 if action == "STOP":
     if photoapp_running:
         print(now, "Stopping python controller")
+        killapp(photoapp_out)
         os.kill(int(photoapp_out), 15)
 elif action == "START":
+    if restart:
+        killapp(photoapp_out)
+        photoapp_running = False
     if not photoapp_running:
         print(now, "Starting python controller")
         print(now, "CMD", PHOTOAPP_CMD)
